requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==3.2.16
djangorestframework==3.15.1
djongo==1.3.6
pymongo==3.12.1
pytz==2024.1
sqlparse==0.2.4
typing-extensions==4.11.0
